Generate skeleton of folders and files for redux
