+++ client/app/App.jsx
@@ -1,0 +1,18 @@
+import React from 'react';
+
+class App extends React.Component {
+
+  constructor(props) {
+    super(props);
+  }
+
+  render () {
+    return (
+      <div>
+        {this.props.children}
+      </div>
+    )
+  }
+}
+
+module.exports = App;
